requirements.txt
langchain>=0.2.0
langchain-community>=0.0.1
langchain-openai
langchain-exa
langgraph
fastapi
fastapi[all]
uvicorn[standard]
python-dotenv
openai
tiktoken
duckduckgo-search
